backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
SQLAlchemy==2.0.36

# AI / Vision
openai==1.61.0
anthropic>=0.44.0

# Image processing
Pillow>=11.1.0
opencv-python-headless>=4.10.0.84

# QR Code reading
pyzbar>=0.1.9

# Text-to-Speech
gTTS>=2.5.4

# Speech-to-Text (fallback)
SpeechRecognition==3.12.0

# Environment variables
python-dotenv==1.0.1